+++ cli/create.go
@@ -1,6 +1,8 @@
 package cli
 
 import (
+	"unicode/utf8"
+
 	"github.com/mlange-42/track/core"
 	"github.com/mlange-42/track/out"
 	"github.com/mlange-42/track/util"
@@ -27,6 +29,7 @@ func createCommand(t *core.Track) *cobra.Command {
 func createProjectCommand(t *core.Track) *cobra.Command {
 	var parent string
 	var color uint8
+	var symbol string
 
 	createProject := &cobra.Command{
 		Use:     "project PROJECT",
@@ -36,10 +39,19 @@ func createProjectCommand(t *core.Track) *cobra.Command {
 		Run: func(cmd *cobra.Command, args []string) {
 			name := args[0]
 
+			if !cmd.Flags().Changed("symbol") {
+				symbol = string([]rune(name)[0])
+			}
+			if utf8.RuneCountInString(symbol) != 1 {
+				out.Err("failed to create project: --symbol must be a single character")
+				return
+			}
+
 			project := core.Project{
 				Name:   name,
 				Parent: parent,
 				Color:  color,
+				Symbol: symbol,
 			}
 
 			if err := t.CheckParents(project); err != nil {
@@ -58,6 +70,7 @@ func createProjectCommand(t *core.Track) *cobra.Command {
 
 	createProject.Flags().StringVarP(&parent, "parent", "p", "", "Parent project of this project")
 	createProject.Flags().Uint8VarP(&color, "color", "c", 0, "Background color for the project, as color index 0..256.\nSee: $ track list colors")
+	createProject.Flags().StringVarP(&symbol, "symbol", "s", "", "Symbol for the project. Defaults to the first letter of the name")
 
 	return createProject
 }