+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Week report with vertical day columns and projects denoted by color and initial letter: `track report week [DATE]` (#35)
 * Auto-scale day and week report width based on terminal width (#35)
 * Projects can be archived (and un-archived) to exclude them from reports etc. (#39)
+* Project colors and symbols are shown in reports and lists (#40)
 
 ### Bug fixes